# docs/swagger.yaml
definitions:
  model.Users:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        description: Unique column
        type: string
      id:
        description: Primary key
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  user.CreateUserRequest:
    properties:
      email:
        default: john@example.com
        example: john@example.com
        type: string
      name:
        default: John Doe
        example: John Doe
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      age:
        default: 30
        example: 30
        type: integer
      email:
        default: updated@example.com
        example: updated@example.com
        type: string
      name:
        default: John Updated
        example: John Updated
        type: string
    type: object
info:
  contact: {}
paths:
  /api/users:
    get:
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Users'
            type: array
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Users Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
      summary: Create user
      tags:
      - users
  /api/users/{user}:
    delete:
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
      summary: Delete user
      tags:
      - users
    get:
      description: Show a user
      parameters:
      - description: User ID
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
      summary: Show user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user
      parameters:
      - description: User ID
        in: path
        name: user
        required: true
        type: string
      - description: Users Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
      summary: Update user
      tags:
      - users
swagger: "2.0"